refactor(tui): split BLOCKS-view applier into submodules
